refactor: canvas options in separate component
